requirements.txt
Flask==3.1.1
line-bot-sdk==3.17.1
python-dotenv==1.1.1
google-genai==1.28.0
notion-client==2.4.0
requests==2.31.0
newspaper3k==0.2.8
nltk==3.9.2
lxml==6.0.0
gunicorn==23.0.0
firebase-admin==7.1.0
cloudscraper==1.2.71
